Move logging tests to own crate
