+++ Cargo.toml
@@ -57,6 +57,7 @@ members = [
 	"client/telemetry",
 	"client/tracing",
 	"client/tracing/proc-macro",
+	"client/tracing/tests",
 	"client/transaction-pool",
 	"client/transaction-pool/graph",
 	"frame/assets",
