+++ client/tracing/Cargo.toml
@@ -13,14 +13,14 @@ readme = "README.md"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-ansi_term = "0.12.1"
+ansi_term = "0.12.1" #needed?
 atty = "0.2.13"
 erased-serde = "0.3.9"
 lazy_static = "1.4.0"
-log = { version = "0.4.8" }
+log = { version = "0.4.8" } # needed?
 once_cell = "1.4.1"
 parking_lot = "0.11.1"
-regex = "1.4.2"
+regex = "1.4.2" # needed?
 rustc-hash = "1.1.0"
 serde = "1.0.101"
 serde_json = "1.0.41"
