+++ client/tracing/tests/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "tracing-integration-tests"
+version = "0.1.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+description = "Integration tests for sc-tracing"
+edition = "2018"
+license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+
+[dev-dependencies]
+ansi_term = "0.12.1"
+log = { version = "0.4.14" }
+regex = "1.4.3"
+tracing = "0.1.22"
+tracing-core = "0.1.17"
+tracing-log = "0.1.1"
+tracing-subscriber = "0.2.15"
+sp-tracing = { version = "2.0.0", path = "../../../primitives/tracing" }
+sc-tracing = { version = "2.0.1", path = "../" }
